+++ lib/converter/pdf_converter.dart
@@ -5,7 +5,11 @@ import 'dart:ui';
 import 'package:dart_quill_delta/dart_quill_delta.dart';
 import 'package:flutter_quill_delta_easy_parser/flutter_quill_delta_easy_parser.dart'
     as ep;
+import 'package:flutter_quill_to_pdf/core/request/font_family_request.dart';
+import 'package:flutter_quill_to_pdf/core/response/font_family_response.dart';
 import 'package:flutter_quill_to_pdf/utils/extensions.dart';
+import 'package:flutter_quill_to_pdf/utils/typedefs.dart';
+import 'package:meta/meta.dart';
 import 'package:pdf/pdf.dart' show PdfColor;
 import 'package:pdf/widgets.dart' as pw;
 import 'package:flutter_quill_to_pdf/flutter_quill_to_pdf.dart' as qpdf;
@@ -30,17 +34,9 @@ class PDFConverter {
   ///[CustomPDFWidget] allow devs to use builders to create custom widgets
   final List<qpdf.CustomWidget> customBuilders;
 
-  ///A simple [request] font when converter detect a font
-  final Future<pw.Font> Function(String)? onRequestFont;
-
-  ///A simple [request] font when converter detect a font
-  final Future<pw.Font> Function(String)? onRequestBoldFont;
-
-  ///A simple [request] font when converter detect a font
-  final Future<pw.Font> Function(String)? onRequestItalicFont;
-
-  ///A simple [request] font when converter detect a font
-  final Future<pw.Font> Function(String)? onRequestBoldItalicFont;
+  ///A font when converter detect a font
+  final FontFamilyResponse Function(FontFamilyRequest familyRequest)?
+      onRequestFontFamily;
 
   ///If you need to [customize] the [theme] of the [pdf document], override this param
   final pw.ThemeData? themeData;
@@ -105,21 +101,22 @@ class PDFConverter {
   final qpdf.PDFWidgetBuilder<List<pw.InlineSpan>, pw.Widget>?
       onDetectBlockquote;
 
-  final Future<List<pw.Font>?> Function(String)? onRequestFallbackFont;
   late final List<pw.Font> globalFontsFallbacks;
+
+  /// [isWeb] is used to know is the current platform is web since the way of the fetch images files
+  /// is different from the other platforms 
+  @experimental
+  final bool isWeb;
 
   PDFConverter({
     required this.pageFormat,
     required this.document,
+    this.isWeb = false,
     this.textDirection = TextDirection.ltr,
     this.frontMatterDelta,
     this.backMatterDelta,
     this.customBuilders = const <qpdf.CustomWidget>[],
-    this.onRequestBoldFont,
-    this.onRequestBoldItalicFont,
-    this.onRequestFallbackFont,
-    this.onRequestFont,
-    this.onRequestItalicFont,
+    this.onRequestFontFamily,
     required List<pw.Font> fallbacks,
     this.blockQuotePaddingLeft,
     this.blockQuotePaddingRight,
@@ -168,83 +165,22 @@ class PDFConverter {
     ];
   }
 
-  // create a pw.Widget that can can be used to draw on a PDF
-  Future<pw.Widget?> generateWidget({
-    qpdf.DeltaAttributesOptions? deltaOptionalAttr,
-    double? maxWidth,
-    double? maxHeight,
-    bool overrideAttributesPassedByUser = false,
-    void Function(dynamic error)? onException,
-    bool shouldProcessDeltas = true,
-  }) async {
-    deltaOptionalAttr ??= qpdf.DeltaAttributesOptions.common();
-    final qpdf.Converter<Delta, pw.Document> converter = qpdf.PdfService(
-      pageFormat: pageFormat,
-      textDirection: textDirection.toPdf(),
-      fonts: globalFontsFallbacks,
-      onRequestBoldFont: onRequestBoldFont,
-      onRequestBothFont: onRequestBoldItalicFont,
-      customTheme: themeData,
-      customBuilders: customBuilders,
-      blockQuoteBackgroundColor: blockQuoteBackgroundColor,
-      blockQuoteDividerColor: blockQuoteDividerColor,
-      codeBlockBackgroundColor: codeBlockBackgroundColor,
-      codeBlockFont: codeBlockFont,
-      codeBlockNumLinesTextStyle: codeBlockNumLinesTextStyle,
-      codeBlockTextStyle: codeBlockTextStyle,
-      blockQuoteTextStyle: blockQuoteTextStyle,
-      onRequestFallbacks: onRequestFallbackFont,
-      onDetectAlignedParagraph: onDetectAlignedParagraph,
-      onDetectCommonText: onDetectCommonText,
-      onDetectBlockquote: onDetectBlockquote,
-      onDetectCodeBlock: onDetectCodeBlock,
-      blockQuotePaddingLeft: blockQuotePaddingLeft,
-      blockQuotePaddingRight: blockQuotePaddingRight,
-      blockQuotethicknessDividerColor: blockQuotethicknessDividerColor,
-      onDetectHeaderBlock: onDetectHeaderBlock,
-      onDetectImageBlock: onDetectImageBlock,
-      onDetectInlineRichTextStyles: onDetectInlineRichTextStyles,
-      onDetectLink: onDetectLink,
-      onDetectList: onDetectList,
-      onRequestFont: onRequestFont,
-      backM: !shouldProcessDeltas
-          ? backMatterDelta
-          : processDelta(backMatterDelta, deltaOptionalAttr,
-              overrideAttributesPassedByUser),
-      frontM: !shouldProcessDeltas
-          ? frontMatterDelta
-          : processDelta(frontMatterDelta, deltaOptionalAttr,
-              overrideAttributesPassedByUser),
-      onRequestItalicFont: onRequestItalicFont,
-      document: !shouldProcessDeltas
-          ? document
-          : processDelta(
-              document, deltaOptionalAttr, overrideAttributesPassedByUser)!,
-    );
-    try {
-      return await converter.generateWidget(
-          maxWidth: maxWidth, maxHeight: maxHeight);
-    } catch (e) {
-      onException?.call(e);
-      rethrow;
-    }
-  }
-
   ///Creates the PDF document an return this one
   Future<pw.Document?> createDocument({
     qpdf.DeltaAttributesOptions? deltaOptionalAttr,
     bool overrideAttributesPassedByUser = false,
     void Function(dynamic error)? onException,
+    PageBuilder? pageBuilder,
     bool shouldProcessDeltas = true,
   }) async {
     deltaOptionalAttr ??= qpdf.DeltaAttributesOptions.common();
     final qpdf.Converter<Delta, pw.Document> converter = qpdf.PdfService(
       pageFormat: pageFormat,
       fonts: globalFontsFallbacks,
-      onRequestBoldFont: onRequestBoldFont,
-      onRequestBothFont: onRequestBoldItalicFont,
       customTheme: themeData,
       textDirection: textDirection.toPdf(),
+      pageBuilder: pageBuilder,
+      isWeb: isWeb,
       customBuilders: customBuilders,
       blockQuoteBackgroundColor: blockQuoteBackgroundColor,
       blockQuoteDividerColor: blockQuoteDividerColor,
@@ -253,7 +189,6 @@ class PDFConverter {
       codeBlockNumLinesTextStyle: codeBlockNumLinesTextStyle,
       codeBlockTextStyle: codeBlockTextStyle,
       blockQuoteTextStyle: blockQuoteTextStyle,
-      onRequestFallbacks: onRequestFallbackFont,
       onDetectAlignedParagraph: onDetectAlignedParagraph,
       onDetectCommonText: onDetectCommonText,
       onDetectBlockquote: onDetectBlockquote,
@@ -266,7 +201,7 @@ class PDFConverter {
       onDetectInlineRichTextStyles: onDetectInlineRichTextStyles,
       onDetectLink: onDetectLink,
       onDetectList: onDetectList,
-      onRequestFont: onRequestFont,
+      onRequestFontFamily: onRequestFontFamily,
       backM: !shouldProcessDeltas
           ? backMatterDelta
           : processDelta(backMatterDelta, deltaOptionalAttr,
@@ -275,7 +210,6 @@ class PDFConverter {
           ? frontMatterDelta
           : processDelta(frontMatterDelta, deltaOptionalAttr,
               overrideAttributesPassedByUser),
-      onRequestItalicFont: onRequestItalicFont,
       document: !shouldProcessDeltas
           ? document
           : processDelta(
@@ -291,16 +225,15 @@ class PDFConverter {
 
   /// This Create the PDF document and write it to storage path
   /// This implementation can throw PathNotFoundException or exceptions based in Storage capabilities
-  ///
-  /// [isWeb] is used to know is the current platform is web since the way of the how is saved PDF file
-  /// is different from the common on mobile devices or Desktop
   Future<void> createDocumentFile({
     required String path,
     void Function(dynamic error)? onException,
     void Function([Object? data])? onSucessWrite,
     qpdf.DeltaAttributesOptions? deltaOptionalAttr,
     bool overrideAttributesPassedByUser = false,
+    PageBuilder? pageBuilder,
     bool shouldProcessDeltas = true,
+    @Deprecated('Use isWeb global variable from PDFConverter instead')
     bool isWeb = false,
   }) async {
     deltaOptionalAttr ??= qpdf.DeltaAttributesOptions.common();
@@ -308,10 +241,9 @@ class PDFConverter {
       pageFormat: pageFormat,
       fonts: globalFontsFallbacks,
       customBuilders: customBuilders,
-      onRequestBoldFont: onRequestBoldFont,
-      onRequestBothFont: onRequestBoldItalicFont,
-      onRequestFallbacks: onRequestFallbackFont,
-      onRequestFont: onRequestFont,
+      pageBuilder: pageBuilder,
+      isWeb: this.isWeb,
+      onRequestFontFamily: onRequestFontFamily,
       onDetectAlignedParagraph: onDetectAlignedParagraph,
       onDetectCommonText: onDetectCommonText,
       customTheme: themeData,
@@ -341,7 +273,6 @@ class PDFConverter {
           ? frontMatterDelta
           : processDelta(frontMatterDelta, deltaOptionalAttr,
               overrideAttributesPassedByUser),
-      onRequestItalicFont: onRequestItalicFont,
       document: !shouldProcessDeltas
           ? document
           : processDelta(
@@ -356,7 +287,7 @@ class PDFConverter {
           ..href =
               "data:application/octet-stream;charset=utf-16le;base64,${base64.encode(fileInts)}"
           ..setAttribute(
-              "download", "${DateTime.now().millisecondsSinceEpoch}.pdf")
+              "download", File(path).uri.pathSegments.last)
           ..click();
         onSucessWrite?.call('');
         return;
@@ -381,4 +312,64 @@ class PDFConverter {
         .withBrackets;
     return Delta.fromJson(jsonDecode(json));
   }
+
+  /// Return a container with the widgets generated from the Document passed 
+  Future<pw.Widget?> generateWidget({
+    qpdf.DeltaAttributesOptions? deltaOptionalAttr,
+    double? maxWidth,
+    double? maxHeight,
+    bool overrideAttributesPassedByUser = false,
+    void Function(dynamic error)? onException,
+    bool shouldProcessDeltas = true,
+  }) async {
+    deltaOptionalAttr ??= qpdf.DeltaAttributesOptions.common();
+    final qpdf.Converter<Delta, pw.Document> converter = qpdf.PdfService(
+      pageFormat: pageFormat,
+      textDirection: textDirection.toPdf(),
+      onRequestFontFamily: onRequestFontFamily,
+      isWeb: isWeb,
+      fonts: globalFontsFallbacks,
+      customTheme: themeData,
+      customBuilders: customBuilders,
+      blockQuoteBackgroundColor: blockQuoteBackgroundColor,
+      blockQuoteDividerColor: blockQuoteDividerColor,
+      codeBlockBackgroundColor: codeBlockBackgroundColor,
+      codeBlockFont: codeBlockFont,
+      codeBlockNumLinesTextStyle: codeBlockNumLinesTextStyle,
+      codeBlockTextStyle: codeBlockTextStyle,
+      blockQuoteTextStyle: blockQuoteTextStyle,
+      onDetectAlignedParagraph: onDetectAlignedParagraph,
+      onDetectCommonText: onDetectCommonText,
+      onDetectBlockquote: onDetectBlockquote,
+      onDetectCodeBlock: onDetectCodeBlock,
+      blockQuotePaddingLeft: blockQuotePaddingLeft,
+      blockQuotePaddingRight: blockQuotePaddingRight,
+      blockQuotethicknessDividerColor: blockQuotethicknessDividerColor,
+      onDetectHeaderBlock: onDetectHeaderBlock,
+      onDetectImageBlock: onDetectImageBlock,
+      onDetectInlineRichTextStyles: onDetectInlineRichTextStyles,
+      onDetectLink: onDetectLink,
+      onDetectList: onDetectList,
+      backM: !shouldProcessDeltas
+          ? backMatterDelta
+          : processDelta(backMatterDelta, deltaOptionalAttr,
+              overrideAttributesPassedByUser),
+      frontM: !shouldProcessDeltas
+          ? frontMatterDelta
+          : processDelta(frontMatterDelta, deltaOptionalAttr,
+              overrideAttributesPassedByUser),
+      document: !shouldProcessDeltas
+          ? document
+          : processDelta(
+              document, deltaOptionalAttr, overrideAttributesPassedByUser)!,
+    );
+    try {
+      return await converter.generateWidget(
+          maxWidth: maxWidth, maxHeight: maxHeight);
+    } catch (e) {
+      onException?.call(e);
+      rethrow;
+    }
+  }
+
 }